# HealthApp.Angular/swagger-api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HealthApp API",
    "version": "v1"
  },
  "paths": {
...
    "/api/PatientApi/paged": {
      "get": {
        "tags": [
          "PatientApi"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/PatientApi/{id}": {
      "get": {
        "tags": [
          "PatientApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/PatientApi": {
      "post": {
        "tags": [
          "PatientApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/PatientApi/me": {
      "get": {
        "tags": [
          "PatientApi"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "PatientApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PatientDto": {
        "required": [
          "dateOfBirth",
          "email",
          "fullName",
          "gender",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "patientId": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "maxLength": 200,
            "minLength": 3,
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "minLength": 1,
            "pattern": "^(Male|Female|Other)$",
            "type": "string"
          },
          "phoneNumber": {
            "minLength": 1,
            "pattern": "^[0-9]{10}$",
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "insuranceId": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "identityUserId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Enter JWT token only. Do not type Bearer manually.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearer": [ ]
    }
  ],
...
}